pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "13.0.0"
requests = "2.20.0"
google-api-python-client = "1.7.9"
oauth2client = "4.1.3"
dialogflow = "0.6.0"
apiai = "^1.2.3"
google-cloud-dialogflow = "^2.28.2"
logger = "^1.4"
pydantic = "^2.6.3"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

